backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
httpx==0.28.1
aiohttp==3.9.3
pydantic==2.6.1
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
xgboost==2.0.3
scipy==1.10.1
joblib==1.2.0
matplotlib==3.7.1
seaborn==0.12.2
requests==2.31.0
python-multipart==0.0.6
SQLAlchemy==2.0.25
aiosqlite==0.19.0
alembic==1.12.1
torch==2.2.0
nba_api==1.4.1
gunicorn==21.2.0
websockets==12.0
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
beautifulsoup4==4.12.3
